+++ controllers/ec.controller.js
@@ -1,16 +1,13 @@
 // declarar variable a exportar
 const ec = {};
-
 // Requerimos pool de base de datos si es necesario
 const pool = require("../models/db");
 const { getUserDataByToken } = require("../middlewares/auth");
 const { Polly } = require("aws-sdk");
-
 ec.main = (req, res) => {
   const usuario = getUserDataByToken(req.cookies.token);
   res.render("ec/formulario", usuario);
 };
-
 ec.administrador = async (req, res) => {
   const usuario = getUserDataByToken(req.cookies.token);
   try {
@@ -29,7 +26,6 @@ ec.administrador = async (req, res) => {
     res.json({ status: false, error }).status(400);
   }
 };
-
 ec.carreras = async (req, res) => {
   try {
     const carrera = await pool.query("SELECT * FROM tb_ec_carrera");
@@ -39,7 +35,6 @@ ec.carreras = async (req, res) => {
     res.json({ status: false, error }).status(400);
   }
 };
-
 ec.grupos = async (req, res) => {
   try {
     const { carrera } = req.params;
@@ -53,7 +48,6 @@ ec.grupos = async (req, res) => {
     res.json({ status: false, error }).status(400);
   }
 };
-
 ec.form = async (req, res) => {
   try {
     const {
@@ -65,22 +59,18 @@ ec.form = async (req, res) => {
       Nombres,
       Apellidos,
     } = req.body;
-
     const statment = `INSERT INTO tb_ec_alumno( carnet, Nombres, Apellidos ,id_grupo, json1, json2, json3) VALUES(  ' ${carnet} ', '${Nombres}' ,'${Apellidos}' ,${grupo}, '${JSON.stringify(
       global_json1
     )} ' , '${JSON.stringify(global_json2)} ' , ' ${JSON.stringify(
       global_json3
     )} ')    `;
-
     await pool.query(statment);
-
     res.json({ status: true });
   } catch (error) {
     console.log(error);
     res.status(400).json(error);
   }
 };
-
 ec.addCarrera = async (req, res) => {
   const { nombrecarrera } = req.body;
   try {
@@ -93,12 +83,11 @@ ec.addCarrera = async (req, res) => {
     res.status(400).json(error);
   }
 };
-
 ec.addGrupo = async (req, res) => {
   const { nombregrupo, id_carrera } = req.body;
   const year = new Date().getFullYear();
   const lastday = function (m, n) {
-    return `${new Date(year, m, 0).getDate()}/${m}/${year+n}`;
+    return `${new Date(year, m, 0).getDate()}/${m}/${year + n}`;
   };
   try {
     const {
@@ -141,8 +130,8 @@ ec.addGrupo = async (req, res) => {
     });
     const promesasUnidades = [];
     datosModeloOrdenado.forEach(async (element) => {
-      let FechaFin = lastday(`${element.fechaFin}` ,0);
-      if(element.fechaInicio >= element.fechaFin){
+      let FechaFin = lastday(`${element.fechaFin}`, 0);
+      if (element.fechaInicio >= element.fechaFin) {
         FechaFin = lastday(`${element.fechaFin}`, 1);
       }
       let FechaInicio = `01/${element.fechaInicio}/${year}`;
@@ -166,13 +155,12 @@ ec.addGrupo = async (req, res) => {
     res.status(400).json(error);
   }
 };
-
 ec.addModelo = async (req, res) => {
-  const { Nombre, Inicio, Fin, idCarrera } = req.body;
+  const { Nombre, Inicio, Fin, idCarrera, horas } = req.body;
   try {
     await pool.query(
-      "INSERT INTO tb_ec_modulos(Nombre, fechaInicio, fechaFin, Estado, isModel, idCarrera) VALUES (?, ? ,? , 1 , 1 ,? )",
-      [Nombre, Inicio, Fin, idCarrera]
+      "INSERT INTO tb_ec_modulos(Nombre, fechaInicio, fechaFin, horas,  Estado, isModel, idCarrera) VALUES (?, ? ,? , ?, 1 , 1 ,? )",
+      [Nombre, Inicio, Fin, horas, idCarrera]
     );
     res.json({ status: true });
   } catch (error) {
@@ -181,11 +169,11 @@ ec.addModelo = async (req, res) => {
   }
 };
 ec.addUnidad = async (req, res) => {
-  const { Nombre, idModulo, idCarrera } = req.body;
+  const { Nombre, idModulo, idCarrera , horas } = req.body;
   try {
     await pool.query(
-      "INSERT INTO tb_ec_unidades(Nombre, isModel, idModulo, idCarrera) VALUES (?, 1 ,? ,?)",
-      [Nombre, idModulo, idCarrera]
+      "INSERT INTO tb_ec_unidades(Nombre, horas , isModel, idModulo, idCarrera) VALUES (?, ?, 1 ,? ,?)",
+      [Nombre, horas, idModulo, idCarrera]
     );
     res.json({ status: true });
   } catch (error) {
@@ -214,10 +202,11 @@ ec.deleteUnidad = async (req, res) => {
   }
 };
 ec.editUnidad = async (req, res) => {
-  const { idUnidad, unidad } = req.body;
+  const { idUnidad, unidad  , horas} = req.body;
   try {
-    await pool.query("UPDATE tb_ec_unidades SET Nombre=? WHERE id=?  ", [
+    await pool.query("UPDATE tb_ec_unidades SET Nombre=? , horas=? WHERE id=?  ", [
       unidad,
+      horas,
       idUnidad,
     ]);
     res.json({ status: true });
@@ -227,11 +216,11 @@ ec.editUnidad = async (req, res) => {
   }
 };
 ec.editModulo = async (req, res) => {
-  const { idModulo, Nombre, Inicio, Fin } = req.body;
+  const { idModulo, Nombre, Inicio, Fin , horas } = req.body;
   try {
     await pool.query(
-      "UPDATE tb_ec_modulos SET Nombre=?  , fechaInicio=?, fechaFin = ?  WHERE id=?  ",
-      [Nombre, Inicio, Fin, idModulo]
+      "UPDATE tb_ec_modulos SET Nombre=?  , fechaInicio=?, fechaFin = ?, horas = ?  WHERE id=?  ",
+      [Nombre, Inicio, Fin, horas, idModulo]
     );
     res.json({ status: true });
   } catch (error) {
@@ -257,11 +246,11 @@ ec.administradorModelo = async (req, res) => {
       idCarrera
     );
     let modulos = pool.query(
-      "SELECT id , Nombre, fechaInicio, fechaFin FROM tb_ec_modulos WHERE isModel = 1 AND  idCarrera = ?",
+      "SELECT id , Nombre, fechaInicio, fechaFin, horas FROM tb_ec_modulos WHERE isModel = 1 AND  idCarrera = ?",
       idCarrera
     );
     let unidades = pool.query(
-      "SELECT id , Nombre, idModulo FROM tb_ec_unidades WHERE isModel = 1 AND  idCarrera = ?",
+      "SELECT id , Nombre, idModulo , horas FROM tb_ec_unidades WHERE isModel = 1 AND  idCarrera = ?",
       idCarrera
     );
     const query = await Promise.all([carrera, modulos, unidades]);
@@ -285,20 +274,25 @@ ec.administradorModelo = async (req, res) => {
     const datosModeloOrdenado = [];
     modulos.forEach((element, id) => {
       const arrModelo = [];
-      const arrModeloIds = [];
       const obj = {};
+      let totalHorasUnidades = 0;
       unidades.forEach((unidad) => {
         if (unidad.idModulo == element.id) {
-          arrModelo.push(unidad.Nombre);
-          arrModeloIds.push(unidad.id);
+          let obj = {};
+          obj["UnidadName"] = unidad.Nombre;
+          obj["UnidadID"] = unidad.id;
+          obj["horas"] = unidad.horas;
+          arrModelo.push(obj);
+          totalHorasUnidades = totalHorasUnidades+unidad.horas
         }
       });
       obj["modelo"] = element.Nombre;
       obj["unidades"] = arrModelo;
-      obj["unidadesID"] = arrModeloIds;
       obj["idModelo"] = element.id;
       obj["fechaInicio"] = getMonth(element.fechaInicio);
       obj["fechaFin"] = getMonth(element.fechaFin);
+      obj["horas"] = element.horas;
+      obj["totalHorasUnidades"] = totalHorasUnidades;
       datosModeloOrdenado[id] = obj;
     });
     res.render("ec/ModeloCurso", {
@@ -307,6 +301,11 @@ ec.administradorModelo = async (req, res) => {
       carrera,
       carreraID: idCarrera,
     });
+    // res.json( {
+    //   datos: datosModeloOrdenado,
+    //   carrera,
+    //   carreraID: idCarrera,
+    // })
   } catch (error) {
     console.log(error);
     res.json({ status: false, error }).status(400);
@@ -328,7 +327,9 @@ ec.administradorCronogramaVigente = async (req, res) => {
       "SELECT tb_ec_unidades.id , tb_ec_unidades.Nombre, tb_ec_unidades.idModulo , tb_ec_unidades.fechaInicio, tb_ec_unidades.fechaFin, tb_ec_unidades.Estado , (SELECT Nombre FROM tb_usuarios WHERE id_usuario = tb_ec_unidades.id_usuario) AS Usuario FROM tb_ec_unidades WHERE isModel = 0 AND  idGrupo = ?",
       idGrupo
     );
-    let profesores = pool.query("SELECT id_usuario, Nombre FROM tb_usuarios WHERE Role = 2");
+    let profesores = pool.query(
+      "SELECT id_usuario, Nombre FROM tb_usuarios WHERE Role = 2"
+    );
     let grupo;
     const query = await Promise.all([carrera, modulos, unidades, profesores]);
     carrera = query[0][0].Carrera;
@@ -364,8 +365,8 @@ ec.administradorCronogramaVigente = async (req, res) => {
       datos: datosModeloOrdenado,
       carrera,
       grupo,
-      carreraID: idGrupo,
-      profesores
+      grupoID: idGrupo,
+      profesores,
     });
   } catch (error) {
     console.log(error);
@@ -373,19 +374,58 @@ ec.administradorCronogramaVigente = async (req, res) => {
   }
 };
 ec.editUnidadVigente = async (req, res) => {
-  const { idUnidad, unidad , inicio, fin, profesor } = req.body;
+  const { idUnidad, unidad, inicio, fin, profesor } = req.body;
   try {
-    await pool.query("UPDATE tb_ec_unidades SET Nombre=?, fechaInicio = ?, fechaFin =? , Estado = 1  , id_usuario=? WHERE id=?  ", [
-      unidad,
-      inicio, 
-      fin,
-      profesor,
-      idUnidad
-    ]);
+    await pool.query(
+      "UPDATE tb_ec_unidades SET Nombre=?, fechaInicio = ?, fechaFin =? , Estado = 1  , id_usuario=? WHERE id=?  ",
+      [unidad, inicio, fin, profesor, idUnidad]
+    );
     res.json({ status: true });
   } catch (error) {
     console.log(error);
     res.status(400).json(error);
   }
 };
+ec.addModeloVigente = async (req, res) => {
+  const { Nombre, Inicio, Fin, idGrupo } = req.body;
+  try {
+    let idCarrera = await pool.query(
+      "SELECT id_carrera FROM tb_ec_grupo WHERE id = ? ",
+      [idGrupo]
+    );
+    idCarrera = idCarrera[0].id_carrera;
+    await pool.query(
+      "INSERT INTO tb_ec_modulos(Nombre, fechaInicio, fechaFin, Estado, isModel, idCarrera, idGrupo) VALUES (?, ? ,? , 1 , 0 , ? ,?)",
+      [Nombre, Inicio, Fin, idCarrera, idGrupo]
+    );
+    res.json({ status: true });
+  } catch (error) {
+    console.log(error);
+    res.status(400).json(error);
+  }
+};
+ec.addUnidadVigente = async (req, res) => {
+  const { Nombre, idModulo, profesor, inicio, fin, idGrupo } = req.body;
+  try {
+    let idCarrera = await pool.query(
+      "SELECT id_carrera FROM tb_ec_grupo WHERE id = ? ",
+      [idGrupo]
+    );
+    idCarrera = idCarrera[0].id_carrera;
+    await pool.query(
+      "INSERT INTO tb_ec_unidades(Nombre, isModel, idModulo, idCarrera , fechaInicio, fechaFin, id_usuario , idGrupo, Estado) VALUES (?, 0 ,? ,?, ? , ?, ?,?, 1)",
+      [Nombre, idModulo, idCarrera, inicio, fin, profesor, idGrupo]
+    );
+    res.json({ status: true });
+  } catch (error) {
+    console.log(error);
+    res.status(400).json(error);
+  }
+};
+
+ec.instructor = async (req, res) => {
+  const usuario = getUserDataByToken(req.cookies.token);
+  res.render("ec/instructores/carreras", usuario);
+};
+
 module.exports = ec;