+++ changelog.md
@@ -3,11 +3,22 @@
 
     ## [Unreleased]
     - Visualización del historial de cursos de participantes y empresas
-    - Enviar correo electrónicos a empresas al poner el comentario de revisión completa
-    - Al ingresar una solicitud enviar correo a los participantes para verificar validez
-    - Terminar funcionalidades del cronograma vigente de empresa centro
+    - Al ingresar una solicitud enviar correo a los participantes para verificar validez (PENDIENTE DE APROBACION)
 
-    ## [1.0.5] - 27/05/2021
+
+    ## [1.0.6] - 01/06/2021
+    ### Added
+    - Finalizados las funciones del cronograma vigente Editar eliminar y agregar
+    - Agregada vista de instructores de empresa centro
+    - Agregada posiblidad de editar instructor en curso 
+    - Agregado comentario desde base de datos
+    - Envio de correo a empresas para notificar aprobación de documentos
+
+    ### Changed
+    - Mejora en las peticiones del gestor de documentos, haciendo las peticiones en una sola promesa
+    - El sistema de comentarios ya no utiliza el localstorage para almacenarse, proviene de base de datos
+
+    ## [1.0.50] - 27/05/2021
     ### Added
     - Actualización de funciones del cronograma vigente
     - Activación de unidades en cronograma vigente